refactor(rest): extract admin storage-compression routes into routes/admin_storage_compression.py
The sweep state cache moves onto RestContext.compression_state. The admin settings router include, which sat inside the extracted region, is kept at its original position.
